+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## [0.0.1] - 07.03.2019.
 
 * initial release.
+
+## [0.0.3] - 10.03.2019
+
+* expose index in the item builder function
+* make deleting items instant when change is needed